refactor: move header component outside app
